check tkb

+++ src/main/java/com/example/demoapi2/controller/LayerClassController.java
@@ -3,6 +3,7 @@ package com.example.demoapi2.controller;
 import com.example.demoapi2.dto.*;
 import com.example.demoapi2.exception.ApiInputException;
 import com.example.demoapi2.model.Account;
+import com.example.demoapi2.model.LayerClass;
 import com.example.demoapi2.service.iAccountService;
 import com.example.demoapi2.service.iLayerClassService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -34,7 +35,7 @@ public class LayerClassController {
             LayerClassResponseDTO layerClassResponse = layerClassService.getLayerClassBySubject(layerClass.getSubjectId());
             return layerClassResponse;
         } else {
-            throw new ApiInputException("FBI Warning!!!");
+            throw new ApiInputException("Bạn không có quyền!");
         }
     }
 
@@ -45,21 +46,47 @@ public class LayerClassController {
             LayerClassResponseDTO layerClassResponse = layerClassService.getLayerClassByAccount(account.getId());
             return layerClassResponse;
         } else {
-            throw new ApiInputException("FBI Warning!!!");
+            throw new ApiInputException("Bạn không có quyền!");
         }
     }
 
     @PostMapping("/layer-class/save")
-    public boolean deleteLayerClassesByAccount(@RequestBody LayerClassRequestDTO layerClassRequest) {
+    public boolean saveLayerClassesByAccount(@RequestBody LayerClassRequestDTO layerClassRequest) {
         if (checkAccount(layerClassRequest.getUsername(), layerClassRequest.getPassword())) {
-            AccountResponseDTO account = accountService.loginByAccount(layerClassRequest.getUsername(), layerClassRequest.getPassword());
-            layerClassService.deleteLayerClassesByAccount(account.getId());
-            for(int i = 0; i < layerClassRequest.getListLayerClassIds().size(); i++){
-                layerClassService.updateById(layerClassRequest.getListLayerClassIds().get(i), account.getId());
+            int count = 0;
+            for (int i = 0; i < layerClassRequest.getListLayerClassIds().size(); i++) {
+                LayerClass layerClass = layerClassService.getLayerClassById(layerClassRequest.getListLayerClassIds().get(i));
+                count += layerClass.getSubject().getStc();
             }
-            return true;
+            int[][] arr = new int[13][9];
+            boolean ok = true;
+            for (int i = 0; i < layerClassRequest.getListLayerClassIds().size(); i++) {
+                LayerClass layerClass = layerClassService.getLayerClassById(layerClassRequest.getListLayerClassIds().get(i));
+                for (int j = 0; j < layerClass.getLessonNumber(); j++) {
+                    if (arr[j + layerClass.getLessonStart()][layerClass.getWeekdayValue()] == 0) {
+                        arr[j + layerClass.getLessonStart()][layerClass.getWeekdayValue()] = 1;
+                        continue;
+                    } else {
+                        ok = false;
+                        break;
+                    }
+                }
+            }
+
+            if (count < 12) {
+                throw new ApiInputException("Không đủ 12 tín chỉ!");
+            } else if (ok == false) {
+                throw new ApiInputException("Trùng thời khoá biểu!");
+            } else {
+                AccountResponseDTO account = accountService.loginByAccount(layerClassRequest.getUsername(), layerClassRequest.getPassword());
+                layerClassService.deleteLayerClassesByAccount(account.getId());
+                for (int i = 0; i < layerClassRequest.getListLayerClassIds().size(); i++) {
+                    layerClassService.updateById(layerClassRequest.getListLayerClassIds().get(i), account.getId());
+                }
+                return true;
+            }
         } else {
-            throw new ApiInputException("FBI Warning!!!");
+            throw new ApiInputException("Bạn không có quyền!");
         }
     }
 }
